requirements.txt
protobuf
python-dotenv==1.0.1
streamlit
streamlit_chat==0.1.1
chromadb
google-generativeai